Close world-native design acceptance gaps
